+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Improved logic for drag handles
 - Added new check "Don't snap selection in text fields" (defaults to "true")
 - Selecton mouse listeners now won't be loaded in page unless user starts to actually select text
+- Added more address markers
 - Various performance optimizations
 
 #3.2.7